Admin: Lakeformation: Fix test names
